% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomarker}
\alias{biomarker}
\title{Biomarker data}
\format{
An object of class \code{data.frame} with 252 rows and 11 columns.
}
\source{
internal
}
\usage{
biomarker
}
\description{
simulated biomarker data
}
\keyword{datasets}
