% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinical}
\alias{clinical}
\title{Clinical data}
\format{
An object of class \code{data.frame} with 652 rows and 12 columns.
}
\source{
internal
}
\usage{
clinical
}
\description{
simulated data from clinical studies
}
\keyword{datasets}
